# public/api-docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "\n\nHeader -> api_key :  \n\n",
    "version": "1.0.0",
    "title": "Ecommerce"
  },
  "host": "laravelandvue.test",
  "basePath": "/api",
...
  "paths": {
...
      },"/getHomeData": {
          "get": {
              "tags": [
                  "Home Page Data"
              ],
              "summary": "Home Page Data",
              "operationId": "getHomeData",
              "produces": [
                  "application/json"
              ],
              "security": [
                  {
                      "APIKeyHeader": []
                  }
              ],
              "parameters": [

              ],
              "responses": {
                  "200": {
                      "description": "Data"
                  }
              }
          }
      },"/getCategoriesData": {
          "get": {
              "tags": [
                  "Home Page Data"
              ],
              "summary": "Home Page Data",
              "operationId": "getCategoriesData",
              "produces": [
                  "application/json"
              ],
              "security": [
                  {
                      "APIKeyHeader": []
                  }
              ],
              "parameters": [

              ],
              "responses": {
                  "200": {
                      "description": "Data"
                  }
              }
          }
      },"/getCategoryData/{slug}": {
...
  },
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "api-key"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ]
}